Fix removal of tagged interface and bridge when multiple bss share them
Currently, if multiple dynamic vlans from different local bss share the bridge
and the tagged vlan interface, only the first instance of struct hostapd_vlan
for this vlanid will have the DVLAN_CLEAN_VLAN flag added.
Thus, when this instance is removed, the tagged vlan interface will be removed
from bridge, thought other bss might still need it.
Similarily, the bridge will be left over, as it does not have zero ports when
the first instance of a struct hostapd_vlan is freed.

1. This patch fixes this by having a global (per process) reference counter for
   dynamic tagged vlan and dynamically created bridge interfaces, so they are
   only removed after all local users are freed.
   (struct hapd_interfaces*)->vlan_priv is used to hold src/ap/vlan_init.c
   global per-process data like drv_priv does; right now this is only used for
   the interface reference counting, but could get extended when needed. Then
   possibly some vlan_global_init / vlan_global_deinit should be added, but
   this is just not required now.

2. Additionally, vlan->configured is checked to avoid reference counter
   decreasing before vlan_newlink increased them.

3. In order to avoid race conditions, vlan_dellink is called explicitly
   after hostapd_vlan_if_remove. Otherwise there would be a short timeframe
   between hostapd_vlan_if_remove and vlan_dellink during which the struct
   hostapd_vlan still exists, so ap_sta_bind_vlan would try to attach stations
   to it.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>