test: verify correct vlan operation in multi-bss multi-vlan case
This adds hwsim test ap_vlan_iface_cleanup_multibss. It connects two stations
in different BSS but the same hostapd process. First both stations are in
VLAN 1, then they get reauthenticated into VLAN 2. Due to the ordering of the
stations moving around, this patch checks that bridge and tagged interface
referencing counting is done globally, such that the tagged interface is not
removed too early and not bridge is left over.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>