nl80211: fix CMD_FRAME used with VLANs
This is required for 802.11r with VLANs used in combination with over_ds,
or the action reply frame won't get out on WiFi.

This is because kernel mac80211 CMD_FRAME uses get_sta instead of
get_sta_bss, so the ifidx given needs to be the of the AP_VLAN iface.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>