nl80211: fix CMD_FRAME used with VLANs
This is required for 802.11r with VLANs used in combination with over_ds,
or the action reply frame won't get out on WiFi.

This is because kernel mac80211 CMD_FRAME uses sta_info_get instead of
sta_info_get_bss, so the ifidx given needs to be the of the AP_VLAN iface.

This could or should be fixed in kernel as well, but in order to be able to
run the hwsim tests right now without rebuilding the kernel or in order to
use FT with VLAN on older kernel, this change is needed.

I'm happy to drop this and only use this as a local workaround until the
kernel change is done.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>