FT: add expiration to PMK-R0 and PMK-R1 cache
IEEE 802.11-2012 says: MSK has a lifetime that limits PMK_R0 and PMK_R1
lifetime. This is currently stored in r0_key_lifetime, but cache entries are
not actually removed.

This patch assumes a default MSK lifetime to be 3600s when
wpa_auth_derive_ptk_ft is called. This matches the default eapol_reauth timeout
and should probably be passed in from eapol state machine with some future
changes. For PSK, things could be handled quite differently, but currently are
not, as pmk_cache can be avoided for FT_PSK with some future changes entirely.

The expiration timeout is then passed form R0KH to R1KH. The R1KH verifies that
the given timeout for sanity, it max not exceed the locally configured
r0_key_lifetime.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>