FT: generate PMK_R0 and PMK_R1 for FT-PSK locally
IEEE 802.11-2012 mandates that a station should be able to connect
initially without ft_pmk_cache filled. So the target AP has the PSK
available and thus the same information as the origin AP.
Therefore neither caching nor communication between the APs with respect
to PMK_R0 or PMK_R1 or VLANs is required if the target AP derives the
required PMKs locally.

This patch introduces the generation of the required PMKs locally for
FT-PSK. Additionally, PMK-R0 is not stored (and thus pushed) for FT-PSK.

So for FT-PSK networks, no configuration of inter-AP communication is
needed anymore.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>