tests: test FT-PSK with VLAN
time.sleep in run_roams is required because the target AP
sets the key once the station was associated. There are
races, when the station processes the (re)association
reply AND the test suite starts FT_DS before the AP
processes its local confirmation and thus
wpa_auth_sm_event(ASSOC_FT). Therefore, the ActionFrame
will be lost, as the AP driver is missing the key.

Since this is this speed is highly synthetic, wait a
few milli seconds before roaming back.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>