FT-over-DS: use macaddr_acl
With FT-over-air, the station sends an authentication request which is
subject to macaddr_acl (RADIUS) in src/ieee802_11.c. But with
FT-over-DS, the authentication request is replaced by an FT-action frame
and FT inter-AP messages, which are not subject to any macaddr_acl logic.

This patch fixes this by adopting macaddr_acl processing for FT over DS.

This ensures session_timeout, radius accounting configuration and alike
even if ft_psk_generate_local is used.

In order to correctly handle this, a new sta flag is introduced.
Therefore, sta->flags is converted to u64. Though, the BIT marcro only
works for <32bit (https://gcc.gnu.org/ml/gcc/1999-09n/msg00860.html).

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>

FIXME: verify accounting is not started with FT, session_timeout and alike
       is already handled by RRB communication.