FT-over-DS: use macaddr_acl
With FT-over-air, the station sends an authentication request which is
subject to macaddr_acl (RADIUS) in src/ieee802_11.c.
But with FT-over-DS, the authentication request is replaced by an
FT-action frame and FT inter-AP messages, which are not subject to any
macaddr_acl logic.

This patch fixes this by adopting macaddr_acl processing for FT over DS.

This ensures session_timeout, radius accounting configuration and alike.

In order to correctly handle this, a new sta flag is introduced.
Therefore, sta->flags is converted to u64. Though, the BIT marcro only
works for <32bit (https://gcc.gnu.org/ml/gcc/1999-09n/msg00860.html).

FIXME: Alternativ approach: FT roaming stations might not need to be
subject to hostapd_allowed_adress when sending FT-over-air authentication
request as the R0KH is expect to have completed this. While FT-over-air
might afford the extra RADIUS RTT as the station is still connected to
the remote AP, FT-over-AIR might want to focus on minimum time without
connection for the STA.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>