test: FT RRB internal delivery to non-WPA BSS
A malcious station could try to do FT-over-DS with a non WPA-enabled BSS. When
this BSS is located in the same hostapd instance, internal RRB delivery will be
used and thus the FT Action Frame will be processed by a non-WPA enabled BSS.
This processing will crash hostapd as hapd->wpa_auth is null. the target BSS
will not listen for these packets and thus not crash.

This testcase checks for this.

If the BSS is located in a different hostapd instance, the target BSS will not
listen for these packets and thus not crash.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>