STA: avoid driver warning when configuring not connected STAs for WPA
At least mac80211 drivers refuse to set VLAN for a station not yet
associated and therefore print errors to console that can be ignored.
In order to avoid confusion from these errors, do not try to configure
station vlan to driver if station is not associated.

It does not fix any real bugs.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>