FT: do not change ANonce during re-sent auth request
The A-Nonce generation is in the authentication request processing that has
strict timing requirements on client side and mac80211 will automatically
resend the request if the AP does not answer within time.

Generating the reply might involve querying RADIUS and PMK-R0-Holder using
RRB as well as FT key derivation and vlan_add_dynamic. This might take
longer than the client expects and thus trigger authentication request
resend.

If the client resends the authentication request not while hostapd is
waiting for RADIUS or RRB pull reply but while it is processing RRB reply
(or shortly blocked for other reasons while RRB reply is already in receive
buffer), hostapd will process the second request after it has send out the
first authentication reply.
For me, this has only happened with address sanitizer enabled.

Processing the second authentication request will overwrite the first
A-Nonce, though the client (wpa_supplicant) will use the first A-Nonce.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>