FT: negative caching when using wilcard r0kh
When a station uses an invalid or offline r0kh_id, requests are always
broadcasted.  In order to avoid this, if r0kh does not reply, a temporary
blacklist entry is added to r0kh_list.

In order to avoid blocking a valid r0kh when a non-existing pmk_r0_name
is requested, r0kh is required to always reply using a NAK. Resend requests
a few times to ensure blacklisting does not happen due to small packet loss.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>

--
v3: merge retransmission into this patch and update comment