FT RRB: add replay protection
This adds a per-BSS sequence counter and adds sequence numbering to FT RRB
packets. The sequence number is checked against r0kh/r1kh sequence number
cache.

Special attention is needed in case the remote AP reboots and thus loses
its state. I prefer it to recover automatically even without synchronized
clocks. Therefore an identifier called dom is generated randomly along the
initial sequence number. If the dom transmitted does not match or the
sequence number is not in the 0x80000000 long range currently expected,
the sender is asked for a fresh confirmation of its currently used sequence
numbers. The packet that triggered this is cached and replayed later.

The TIMESTAMP attribute is not removed from push messages, as this
mechanism does not guarantee freshness in low-rate settings. To fix this,
push messages would probably need a 3-way handshake.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>