FT RRB: add msg replay and msg delay protection
This adds a per-BSS sequence counter and adds sequence numbering to FT RRB
packets. The sequence number is checked against r0kh/r1kh sequence number
cache.

Special attention is needed in case the remote AP reboots and thus loses its
state. I prefer it to recover automatically even without synchronized clocks.
Therefore an identifier called dom is generated randomly along the initial
sequence number. If the dom transmitted does not match or the sequence number
is not in the range currently expected, the sender is asked for a fresh
confirmation of its currently used sequence numbers. The packet that triggered
this is cached and processed again later.

Additionally, in order to ensure freshness, the remote ap includes an
timestamp with its messages. It is then verified that the received
messages are indeed fresh by comparing it to the older timestamps received
and the time elapsed since then.
Therefore FT_RRB_TIMESTAMP is no longer needed.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>